Initial Commit
This is a website that I created as my final project for the SheCodes Basics course. I utilized the skills I learned from the course to build this site with HTML, CSS, and JavaScript. I also researched different hover effects, such as the bold line around the main image and the colors moving across the text towards the bottom of the page.